requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
Django==1.11.1
django-crispy-forms==1.7.0
django-mathjax==0.0.8
gunicorn==19.7.1
html5lib==1.0b10
httplib2==0.10.3
idna==2.6
kombu==4.1.0
more-itertools==3.1.0
numpy==1.12.1
olefile==0.44
pandas==0.20.3
Pillow==4.2.1
psycopg2==2.7.3.2
PyPDF2==1.26.0
python-dateutil==2.6.1
pytz==2017.2
redis==2.10.6
reportlab==3.4.0
requests==2.18.4
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
urllib3==1.22
vine==1.1.4
webencodings==0.5.1
whitenoise==3.3.0
xhtml2pdf==0.2b1
XlsxWriter==1.0.2
djangorestframework==3.8.2

